Add ANALYZE support to compression TAM
To ANALYZE a table using compression TAM, it is necessary to sample
storage blocks across both the non-compressed and compressed
relation. Unfortunately, PostgreSQL expects a relation's storage to be
a set of contiguous blocks which it can iterate over when sampling.

To make this work, it is necessary to give the illusion that the
compression TAM storage is the sum of the non-compressed and
compressed blocks, by, e.g., returning this sum in the
`relation_size()` TAM callback.

However, this is complicated by the fact that the user-visible
relation is also the a regular heap that holds non-compressed data and
returning a relation size that doesn't reflect the actual number of
blocks breaks other code paths, like INSERT.

Thus, this change implements a workaround by only returning the summed
relation size during ANALYZE. However, a better approach for the
future might be to have three relations: one meta-relation that is
user-visible and represents the complete storage, and then two
sub-relations for the actual non-compressed and compressed data,
respectively. Then the sub-relations can return accurate relation
sizes while the meta-relation returns the aggregate.