Fix relation_size() handling
When the TAM's `relation_size()` callback is called, it should return
the combined size of the non-compressed and compressed
relations. However, when scanning non-compressed data, the
`relation_size()` should reflect only the non-compressed data. This
leads to issues because the non-compressed relation and the TAM's
user-visible relation are both the same relation.

This change handles this situation by switching TAM callbacks on the
relation back to HeapAM for the duration of internal calls. Thus, the
"upward" visible callbacks are hyperstore TAM, while the "downward"
visible callbacks are HeapAM.

Switching callbacks like this simplifies a lot of code the previously
tried to handle the relation_size()-issue and makes stats and ANALYZE
work much better.