Add support for VACUUM FULL in hyperstore
VACUUM FULL in Hyperstore works similar to Heap, i.e., a table is
rewritten completely using a heap swap. However, in the case of
Hyperstore, compression happens at the same time and uncompressed data
is moved from the non-compressed relation to the compressed
relation. Instead of one heap swap, two heap swaps are performed
because both the non-compressed and compressed tables are
rewritten. Note, however, that the heap swap on the non-compressed
table works more like a TRUNCATE as it will be empty after the
operation completes.