Fix bug with copying tuples from file
When copying from a file to a hyperstore an Arrow tuple is used both as
source and destination. Since `copyfrom` fills in the virtual part of
the tuple and calls `ExecStoreVirtualTupleSlot` rather than either of
the child slots, the copy method fails since it assumes that not both
child slots are empty.

This is fixed by explicitly calling the virtual tuple slot method when
both child slots are either empty or null.