Relax blocking of constraint creation on compressed hypertables
Allow adding constraints to hypertables that have compression enabled
when they don't have compressed data. Previously, there was a blanket
block of these operations on the hypertable level as soon as
compression settings existed.

By moving checks from the hypertable level to chunks instead, it is
possible to be more permissive when adding check, unique, foreign key,
and other constraints as long as there is no compressed data.