Do not vacuum the hypertable relation
When vacuuming a hypertable, all chunks should be vacuumed but the
hypertable relation itself does not need to be vacuumed.

Running vacuum analyze on a hyperstore hypertable will cause an
assertion failure since the code assumes that a cache entry with a
proper `compressed_relid` can be constructed, which is not possible for
the hypertable relation itself since it does not have a corresponding
compressed relation.

We solve this by creating a stateful scan descriptor that can either be
working on the compressed relation, the non-compressed relation, or
done with the scan and set it to "done" immediately for hypertable
relations.