Use bool array for valid_attrs in arrow slot
The valid_attrs in the arrow tuple table slot is a bitmapset used to
track columns/attributes that are "materialized" in the slot. This
bitmapset is cleared by freeing the set and then reallocating it again
for the next row in an arrow array. The reallocation happens on a
performance-critical "hot path", and has a significant peformance
impact.

The performance is improved by making valid_attrs a bool array
instead, and preallocating the array at slot initialization. Clearing
it is a simple memset(). While a bool array takes a bit more space
than a bitmapset, it has simpler semantics and is always the size of
the number of attributes.