Pass on scankeys in parallel columnar scans
This fixes an issue where scankeys were not applied in parallel scans
due to PG not passing on the scan keys to the underlying table access
method when using the function `table_beginscan_parallel()`.

To test the use of scankeys in parallel scans, a test is added that
uses a filter on a segmentby column (this is, currently, the only case
where scankeys are used instead of quals).